pyproject.toml
[project]
name = "coldcamera"
version = "0.1.1"
description = ""
authors = [
    {name = "deathloveleopards", email = "deathloveleopards@gmail.com"},
    {name = "Serhii Chyrva", email = "serhiichyrva@hotmail.com"}
]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.11, <3.12"
dependencies = [
    "pyside6 (>=6.9.2,<7.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "numpy (==2.2.6)",
    "loguru (>=0.7.3,<0.8.0)",
    "pyqtdarktheme (>=2.1.0,<3.0.0)",
    "moderngl (>=5.12.0,<6.0.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "blend-modes (>=2.2.0,<3.0.0)",
    "pedalboard (>=0.9.22,<0.10.0)"
]


[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.16.0"


[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"


[project.urls]
Homepage = "https://github.com/deathloveleopards/coldcamera"
Repository = "https://github.com/deathloveleopards/coldcamera"
Issues = "https://github.com/deathloveleopards/coldcamera/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["coldcamera", "coldcamera.*"]

[tool.ruff]
line-length = 160
